requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
Flask-Migrate==4.0.5
qrcode==7.4.2
Pillow==10.0.1
python-dotenv==1.0.0
bcrypt==4.0.1
datetime
uuid
requests
gunicorn==21.2.0 